Move and fix mapToLegacyLineName
